+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Log message when setting user before starting the SDK (#4882)
 - Add experimental flag to disable swizzling of `NSData` individually (#4859)
 - Replace calls of `SentryScope.useSpan` with callback to direct span accessor (#4896)
+- Change type of Swift String constants to NSString (#4894)
 
 ### Fixes